add subcommands to tile-pipeline command

+++ aviary/_cli/cli.py
@@ -1,6 +1,7 @@
 from collections.abc import Callable
 from functools import wraps
 from gettext import gettext as _
+from itertools import groupby
 from pathlib import Path
 from typing import Any
 
@@ -27,14 +28,16 @@ from aviary._cli.templates import registry as template_registry
 from aviary._utils.plugins import discover_plugins
 from aviary.pipeline.tile_pipeline import (
     TilePipelineConfig,
-    TilePipelineFactory,
+    _TilePipelineFactory,
 )
 
 # noinspection PyProtectedMember
-from aviary.tile.tile_fetcher import _registry as tile_fetcher_registry
+from aviary.tile.tile_fetcher import _TileFetcherFactory
 
 # noinspection PyProtectedMember
-from aviary.tile.tiles_processor import _registry as tiles_processor_registry
+from aviary.tile.tiles_processor import _TilesProcessorFactory
+
+_PACKAGE = 'aviary'
 
 typer.rich_utils.DEFAULT_STRING = _('default: {}')
 typer.rich_utils.REQUIRED_LONG_STRING = _('required')
@@ -68,7 +71,10 @@ def version_callback(
     value: bool,
 ) -> None:
     if value:
-        print(f'aviary {__version__}')
+        message = (
+            f'[dim green]aviary [/][bold green]{__version__}'
+        )
+        console.print(message)
         raise typer.Exit(0)
 
 
@@ -112,13 +118,47 @@ def handle_exception(
             )
         except click.ClickException:
             raise
+        except typer.Abort:
+            raise
+        except typer.Exit:
+            raise
         except Exception as error:
             if verbose:
                 raise
 
-            console.print(f'[bold bright_red]{type(error).__name__}:[/] {error}')
+            message = (
+                f'[bold bright_red]{type(error).__name__}:[/] {error}'
+            )
+            console.print(message)
             context.exit(1)
     return wrapper
+
+
+@app.command(
+    rich_help_panel='General commands',
+)
+@handle_exception
+def components(
+    plugins_dir_path_option: Path | None = typer.Option(
+        None,
+        '--plugins-dir-path',
+        help='Path to the plugins directory',
+    ),
+    type_options: list[str] | None = typer.Option(
+        None,
+        '--type',
+        '-t',
+        click_type=click.Choice(['tile_fetcher', 'tiles_processor']),  # noqa: B008
+        help='Type of the components',
+    ),
+) -> None:
+    """Show the components."""
+    show_components(
+        title='Components:',
+        plugins_dir_path=plugins_dir_path_option,
+        type_options=type_options,
+        filter_packages=None,
+    )
 
 
 @app.command(
@@ -144,25 +184,26 @@ def github() -> None:
 )
 @handle_exception
 def plugins(
-    plugins_dir_path: Path = typer.Argument(
-        ...,
+    plugins_dir_path_option: Path | None = typer.Option(
+        None,
+        '--plugins-dir-path',
         help='Path to the plugins directory',
+    ),
+    type_options: list[str] | None = typer.Option(
+        None,
+        '--type',
+        '-t',
+        click_type=click.Choice(['tile_fetcher', 'tiles_processor']),  # noqa: B008
+        help='Type of the components',
     ),
 ) -> None:
     """Show the registered plugins."""
-    discover_plugins(plugins_dir_path=plugins_dir_path)
-
-    tile_fetcher_names = list(tile_fetcher_registry.keys())
-    tiles_processor_names = list(tiles_processor_registry.keys())
-
-    print('Registered plugins:')
-    print('  TileFetcher:')
-    for plugin_name in tile_fetcher_names:
-        print(f'    - {plugin_name}')
-
-    print('  TilesProcessor:')
-    for plugin_name in tiles_processor_names:
-        print(f'    - {plugin_name}')
+    show_components(
+        title='Registered plugins:',
+        plugins_dir_path=plugins_dir_path_option,
+        type_options=type_options,
+        filter_packages=lambda package: package != _PACKAGE,
+    )
 
 
 @tile_pipeline_app.command(
@@ -183,12 +224,27 @@ def tile_pipeline_init(
     ),
 ) -> None:
     """Initialize a config file."""
+    if config_path.exists():
+        message = (
+            f'[bold yellow]The config file [/][dim yellow]at {config_path.resolve()}[/][bold yellow] already exists.'
+        )
+        console.print(message)
+        overwrite = typer.confirm('Do you want to overwrite it?')
+
+        if not overwrite:
+            raise typer.Exit(0)
+
     pipeline = 'tile_pipeline'
     key = (pipeline, template_option)
     config = template_registry[key]
 
     with config_path.open('w') as file:
         file.write(config)
+
+    message = (
+        f'[bold green]The config file [/][dim green]at {config_path.resolve()}[/][bold green] has been initialized.'
+    )
+    console.print(message)
 
 
 @tile_pipeline_app.command(
@@ -204,7 +260,7 @@ def tile_pipeline_run(
         None,
         '--set',
         '-s',
-        help='Set configuration fields using key=value format.',
+        help='Configuration fields using key=value format.',
     ),
 ) -> None:
     """Run the tile pipeline."""
@@ -220,7 +276,7 @@ def tile_pipeline_run(
         discover_plugins(plugins_dir_path=plugins_dir_path)
 
     tile_pipeline_config = TilePipelineConfig(**config)
-    tile_pipeline = TilePipelineFactory.create(config=tile_pipeline_config)
+    tile_pipeline = _TilePipelineFactory.create(config=tile_pipeline_config)
     tile_pipeline()
 
 
@@ -237,7 +293,7 @@ def tile_pipeline_validate(
         None,
         '--set',
         '-s',
-        help='Set configuration fields using key=value format.',
+        help='Configuration fields using key=value format.',
     ),
 ) -> None:
     """Validate the config file."""
@@ -253,7 +309,12 @@ def tile_pipeline_validate(
         discover_plugins(plugins_dir_path=plugins_dir_path)
 
     tile_pipeline_config = TilePipelineConfig(**config)
-    _ = TilePipelineFactory.create(config=tile_pipeline_config)
+    _ = _TilePipelineFactory.create(config=tile_pipeline_config)
+
+    message = (
+        f'[bold green]The config file [/][dim green]at {config_path.resolve()}[/][bold green] is valid.'
+    )
+    console.print(message)
 
 
 def parse_config(
@@ -284,6 +345,74 @@ def parse_config(
             sub_config[sub_keys[-1]] = value
 
     return config
+
+
+def show_components(
+    title: str,
+    plugins_dir_path: Path | None = None,
+    type_options: list[str] | None = None,
+    filter_packages: Callable | None = None,
+) -> None:
+    if plugins_dir_path is not None:
+        discover_plugins(plugins_dir_path=plugins_dir_path)
+
+    tile_fetchers = sorted(
+        [
+            registry_entry
+            for registry_entry in _TileFetcherFactory.registry
+            if filter_packages is None or filter_packages(registry_entry[0])
+        ],
+        key=lambda registry_entry: (registry_entry[0], registry_entry[1]),
+    )
+    tiles_processors = sorted(
+        [
+            registry_entry
+            for registry_entry in _TilesProcessorFactory.registry
+            if filter_packages is None or filter_packages(registry_entry[0])
+        ],
+        key=lambda registry_entry: (registry_entry[0], registry_entry[1]),
+    )
+
+    message = (
+        f'[bold green]{title}'
+    )
+    console.print(message)
+
+    if type_options is None or 'tile_fetcher' in type_options:
+        message = (
+            '  [bold green]TileFetcher:'
+        )
+        console.print(message)
+
+        for package, names in groupby(tile_fetchers, key=lambda registry_entry: registry_entry[0]):
+            message = (
+                f'    [green]{package}:'
+            )
+            console.print(message)
+
+            for _, name in names:
+                message = (
+                    f'      - {name}'
+                )
+                console.print(message)
+
+    if type_options is None or 'tiles_processor' in type_options:
+        message = (
+            '  [bold green]TilesProcessor:'
+        )
+        console.print(message)
+
+        for package, names in groupby(tiles_processors, key=lambda registry_entry: registry_entry[0]):
+            message = (
+                f'    [green]{package}:'
+            )
+            console.print(message)
+
+            for _, name in names:
+                message = (
+                    f'      - {name}'
+                )
+                console.print(message)
 
 
 if __name__ == '__main__':
